fix existing tests
